+++ R/endpoints.R
@@ -209,8 +209,10 @@ pub_covid_hosp_facility <- function(
     cli::cli_warn(coercion_msg, class = "epidatr__epirange_week_coercion")
     collection_weeks <- reformat_epirange(collection_weeks, to_type = "day")
     # Single week date.
-  } else if ((test_integerish(collection_weeks) || test_character(collection_weeks)) &&
-    nchar(collection_weeks) == 6) {
+  } else if (
+    (test_integerish(collection_weeks) || test_character(collection_weeks)) &&
+      nchar(collection_weeks) == 6
+  ) {
     cli::cli_warn(coercion_msg, class = "epidatr__single_week_coercion")
     collection_weeks <- parse_api_week(collection_weeks)
   }
@@ -2094,15 +2096,17 @@ pvt_sensors <- function(
 #' pvt_twitter(
 #'   auth = Sys.getenv("SECRET_API_AUTH_TWITTER"),
 #'   locations = "CA",
-#'   epiweeks = epirange(201501, 202001)
+#'   time_type = "week",
+#'   time_values = epirange(201501, 202001)
 #' )
 #' }
 #' @param auth string. Restricted access key (not the same as API key).
 #' @param locations character. Locations to fetch.
 #' @param ... not used for values, forces later arguments to bind by name
-#' @param dates [`timeset`]. Dates to fetch. Mutually exclusive with `epiweeks`.
-#' @param epiweeks [`timeset`]. Epiweeks to fetch. Mutually exclusive with
-#' `dates`.
+#' @param time_type string. The temporal resolution of the data (either "day" or
+#'  "week", depending on signal).
+#' @param time_values [`timeset`]. Dates or epiweeks to fetch. Defaults to all
+#'  ("*") dates.
 #' @param fetch_args [`fetch_args`]. Additional arguments to pass to `fetch()`.
 #' @return [`tibble::tibble`]
 #' @keywords endpoint
@@ -2111,21 +2115,31 @@ pvt_twitter <- function(
     auth,
     locations,
     ...,
-    dates = NULL,
-    epiweeks = NULL,
+    time_type = c("day", "week"),
+    time_values = "*",
     fetch_args = fetch_args_list()) {
   rlang::check_dots_empty()
 
+  time_type <- match.arg(time_type)
+  if (time_type == "day") {
+    dates <- time_values
+    epiweeks <- NULL
+    dates <- get_wildcard_equivalent_dates(dates, "day")
+  } else {
+    dates <- NULL
+    epiweeks <- time_values
+    epiweeks <- get_wildcard_equivalent_dates(epiweeks, "week")
+  }
+
   assert_character_param("auth", auth, len = 1)
   assert_character_param("locations", locations)
+  assert_character_param("time_type", time_type, len = 1)
+  assert_timeset_param("time_values", time_values)
   assert_timeset_param("dates", dates, required = FALSE)
   assert_timeset_param("epiweeks", epiweeks, required = FALSE)
   dates <- parse_timeset_input(dates)
   epiweeks <- parse_timeset_input(epiweeks)
 
-  if (!xor(is.null(dates), is.null(epiweeks))) {
-    stop("exactly one of `dates` and `epiweeks` is required")
-  }
   time_field <- if (!is.null(dates)) {
     create_epidata_field_info("date", "date")
   } else {
@@ -2163,13 +2177,18 @@ pvt_twitter <- function(
 #'
 #' @examples
 #' \dontrun{
-#' pub_wiki(articles = "avian_influenza", epiweeks = epirange(201501, 201601))
+#' pub_wiki(
+#'   articles = "avian_influenza",
+#'   time_type = "week",
+#'   time_values = epirange(201501, 201601)
+#' )
 #' }
 #' @param articles character. Articles to fetch.
 #' @param ... not used for values, forces later arguments to bind by name
-#' @param dates [`timeset`]. Dates to fetch. Mutually exclusive with `epiweeks`.
-#' @param epiweeks [`timeset`]. Epiweeks to fetch. Mutually exclusive with
-#' `dates`.
+#' @param time_type string. The temporal resolution of the data (either "day" or
+#'  "week", depending on signal).
+#' @param time_values [`timeset`]. Dates or epiweeks to fetch. Defaults to all
+#'  ("*") dates.
 #' @param language string. Language to fetch.
 #' @param hours integer. Optionally, the hours to fetch.
 #' @param fetch_args [`fetch_args`]. Additional arguments to pass to `fetch()`.
@@ -2179,14 +2198,27 @@ pvt_twitter <- function(
 pub_wiki <- function(
     articles,
     ...,
-    dates = NULL,
-    epiweeks = NULL,
+    time_type = c("day", "week"),
+    time_values = "*",
     hours = NULL,
     language = "en",
     fetch_args = fetch_args_list()) {
   rlang::check_dots_empty()
 
+  time_type <- match.arg(time_type)
+  if (time_type == "day") {
+    dates <- time_values
+    epiweeks <- NULL
+    dates <- get_wildcard_equivalent_dates(dates, "day")
+  } else {
+    dates <- NULL
+    epiweeks <- time_values
+    epiweeks <- get_wildcard_equivalent_dates(epiweeks, "week")
+  }
+
   assert_character_param("articles", articles)
+  assert_character_param("time_type", time_type, len = 1)
+  assert_timeset_param("time_values", time_values)
   assert_timeset_param("dates", dates, required = FALSE)
   assert_timeset_param("epiweeks", epiweeks, required = FALSE)
   assert_integerish_param("hours", hours, required = FALSE)
@@ -2194,9 +2226,6 @@ pub_wiki <- function(
   dates <- parse_timeset_input(dates)
   epiweeks <- parse_timeset_input(epiweeks)
 
-  if (!xor(is.null(dates), is.null(epiweeks))) {
-    stop("exactly one of `dates` and `epiweeks` is required")
-  }
   time_field <- if (!is.null(dates)) {
     create_epidata_field_info("date", "date")
   } else {